pyproject.toml
[tool.poetry]
name = "scouts-kampvisum-api"
version = "0.1.0"
description = "Provides functionality for setting up a scouts kamp visum"
authors = ["inuits <info@inuits.eu>"]
license = "GPL v2.0"

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.2.2"
django-rest-framework = "^0.1.0"
django-cors-headers = "^3.7.0"
django-sql-middleware = "^0.0.7"
django-filter = "^2.4.0"
mozilla-django-oidc = "^2.0.0"
drf-yasg2 = "^1.19.4"
cffi = "^1.14.6"
PyYAML = "^5.4.1"
environs = "^9.3.2"
future = "^0.18.2"
psycopg2-binary = "^2.9.1"
django-safedelete = "^1.0.0"
django-rest-enumfield = "^0.2.0"
django-storages = "^1.11.1"
boto3 = "^1.18.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
